Windows: adapt memory and mmap
