+++ CONTRIBUTORS
@@ -1,2 +1,3 @@
 Ryota Nakao
 Doi Yusuke
+Artur Martirosyan (@DoumanAsh)
